+++ packages/taro-cli/src/weapp.js
@@ -123,6 +123,33 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
         (node.superClass.type === 'MemberExpression' &&
         node.superClass.object.name === taroImportDefaultName)) {
           needExportDefault = true
+          astPath.traverse({
+            ClassMethod (astPath) {
+              const node = astPath.node
+              if (node.kind === 'constructor') {
+                astPath.traverse({
+                  ExpressionStatement (astPath) {
+                    const node = astPath.node
+                    if (node.expression &&
+                      node.expression.type === 'AssignmentExpression' &&
+                      node.expression.operator === '=') {
+                      const left = node.expression.left
+                      if (left.type === 'MemberExpression' &&
+                        left.object.type === 'ThisExpression' &&
+                        left.property.type === 'Identifier' &&
+                        left.property.name === 'config') {
+                        configObj = traverseObjectNode(node.expression.right)
+                        if (type === PARSE_AST_TYPE.ENTRY) {
+                          appConfig = configObj
+                        }
+                        astPath.remove()
+                      }
+                    }
+                  }
+                })
+              }
+            }
+          })
           if (node.id === null) {
             componentClassName = '_TaroComponentClass'
             astPath.replaceWith(t.classDeclaration(t.identifier(componentClassName), node.superClass, node.body, node.decorators || []))
@@ -385,8 +412,6 @@ function parseAst (type, ast, depComponents, sourceFilePath, filePath) {
                         objArr = convertObjectToAstExpression(obj)
                       }
                       astPath.replaceWith(t.variableDeclaration('const', [t.variableDeclarator(t.identifier(defaultSpecifier), t.objectExpression(objArr))]))
-                    } else {
-                      astPath.remove()
                     }
                   }
                 } else if (Util.REG_FONT.test(valueExtname) || Util.REG_IMAGE.test(valueExtname) || Util.REG_MEDIA.test(valueExtname)) {
@@ -594,13 +619,13 @@ function convertObjectToAstExpression (obj) {
       return t.objectProperty(t.stringLiteral(key), t.booleanLiteral(value))
     }
     if (Array.isArray(value)) {
-      return t.objectProperty(t.stringLiteral(key), convertArrayToAstExpression(value))
+      return t.objectProperty(t.stringLiteral(key), t.arrayExpression(convertArrayToAstExpression(value)))
     }
     if (value == null) {
       return t.objectProperty(t.stringLiteral(key), t.nullLiteral())
     }
     if (typeof value === 'object') {
-      return t.objectProperty(t.stringLiteral(key), convertObjectToAstExpression(value))
+      return t.objectProperty(t.stringLiteral(key), t.objectExpression(convertObjectToAstExpression(value)))
     }
   })
   return objArr
@@ -1326,7 +1351,16 @@ function watchFiles () {
             if (hasbeenBuiltIndex >= 0) {
               hasBeenBuiltComponents.splice(hasbeenBuiltIndex, 1)
             }
-            await buildSingleComponent(filePath)
+            if (isWindows) {
+              await new Promise((resolve, reject) => {
+                setTimeout(async () => {
+                  await buildSingleComponent(filePath)
+                  resolve()
+                }, 300)
+              })
+            } else {
+              await buildSingleComponent(filePath)
+            }
           } else {
             let isImported = false
             for (const key in dependencyTree) {